add information to matcher UI

+++ MovieMatcher/Models/Api/Movie.cs
@@ -29,7 +29,7 @@ namespace MovieMatcher.Models.Api
         public string homepage { get; set; }
         public string tagline { get; set; }
         public int budget { get; set; }
-        public int revenue { get; set; }
+        public ulong revenue { get; set; }
         public string status { get; set; }
         public string media_type { get; set; } = "movie";
         public BelongsToCollection belongs_to_collection { get; set; }
